% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{.CO2_uitstoot}
\alias{.CO2_uitstoot}
\title{Bereken de CO2 uitstoot bij gegeven id, drooglegging, en slootafstand conform de informatie in de tabel 'brondata'.}
\usage{
.CO2_uitstoot(x)
}
\arguments{
\item{x}{Named vector with id, drooglegging, slootafstand.}
}
\value{
CO2 uitstoot; (numeric)
}
\description{
Bereken de CO2 uitstoot bij gegeven id, drooglegging, en slootafstand conform de informatie in de tabel 'brondata'.
}
