% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{.validate_input}
\alias{.validate_input}
\title{Controleer de input voor een berekening met smrs_CO2_uitstoot().}
\usage{
.validate_input(x)
}
\arguments{
\item{x}{Named vector with id, drooglegging, slootafstand}
}
\value{
foutcode (numeric)
}
\description{
Controleer de input voor een berekening met smrs_CO2_uitstoot().
}
