% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{smrs_CO2_uitstoot}
\alias{smrs_CO2_uitstoot}
\title{Bereken vlakdekkend de CO2 uitstoot met de SOMERS methodiek.}
\usage{
smrs_CO2_uitstoot(r)
}
\arguments{
\item{r}{Raster met layers id, drooglegging (m-mv), slootafstand (m) (SpatRaster)
id is een code (getal) waarvan de cijfers achtereenvolgens zijn samengesteld uit:

- archetype (1-8): 'hV','pV','kV','hVz','kVz','V','Vz','aVz'.

- bron (1-3): <- 'grfr', 'ov', 'wn'.

- type_perceel (1-3): 'ref', 'owd', 'dd'. Opmerking: definieer ook 'scenario' bij gebruik van 'type_perceel'.

- winterpeil (1-4): 'gelijk aan zomerpeil', '10 cm beneden zomerpeil', '20 cm beneden zomerpeil'.

- variabele (1-3): 'CO2-uitstoot.min', 'CO2-uitstoot.mediaan', 'CO2-uitstoot.max'.

- kwel_scenario (0-2): 'nvt', 'lichte kwel', 'lichte wegzijging'. Opmerking: Gegevens mbt kwel zijn alleen beschikbaar in de Overijssel dataset ('bron'='ov', code=2).

- scenario (0-2): 'nvt', 'medium_peil', 'hoog_peil'. Opmerking: alleen van toepassing bij drukdrainage ('type_perceel'='dd', code=3).}
}
\value{
CO2 uitstoot (SpatRaster)
}
\description{
Bereken vlakdekkend de CO2 uitstoot met de SOMERS methodiek.
}
\examples{
r <- terra::rast(system.file('extdata', 'example_input.tif', package='somers', mustWork=TRUE))
CO2_uitstoot <- smrs_CO2_uitstoot(r)
}
\seealso{
\link{smrs_validate_input}
}
