% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{smrs_validate_input}
\alias{smrs_validate_input}
\title{Controleer de input voor de functie 'smrs_CO2_uitstoot()'.}
\usage{
smrs_validate_input(r)
}
\arguments{
\item{r}{Raster met layers id, drooglegging (m-mv), slootafstand (m) (SpatRaster)}
}
\value{
foutcodes (SpatRaster):

-  0: geldige invoer.

- -1: ongeldige id.

- -2: drooglegging out of range.

- -3: slootafstand out of range.

- -4: ongeldige raster layer naam/namen.

- -5: onvoldoende gegevens (1 of meer layers bevatten NA).
}
\description{
Controleer de input voor de functie 'smrs_CO2_uitstoot()'.
}
\examples{
r <- terra::rast(system.file('extdata', 'example_input.tif', package='somers', mustWork=TRUE))
foutcodes <- smrs_validate_input(r)
}
\seealso{
\link{smrs_CO2_uitstoot}
}
