% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somers.r
\docType{package}
\name{somers}
\alias{somers}
\title{somers: De effecten van maatregelen op de CO2 -uitstoot in veenweidegebieden.}
\description{
Achtergrond documentatie:
}
\details{
\href{https://www.nobveenweiden.nl/wp-content/uploads/2022/12/SOMERS-1.0-hoofdrapport-2022-v4.0.pdf}{Subsurface Organic Matter Emission Registration System (SOMERS)}

Functies:

\code{\link{smrs_CO2_uitstoot}}

\code{\link{smrs_validate_input}}
}
